Aggregate the elementary indicators on the provided data matrix

aggregate aggregates the set of elementary indicators through the selected method and computes the composite according to the specified set of weights.

Usage

aggregate(data, method = "linear", w)

Arguments

data

data matrix with the set of normalised elementary indicators (without missing values).

method

aggregation method. Possible choices: "linear" (default) and "non-linear". See Details.

w

vector of weights, as returned by get_weights().

Value

vector of composite indicator values for each target unit in data.

Details

The choice of the aggregation method heavily depends on the degree of compensability or substitutability of the elementary indicators. A compensatory approach requires the use of linear functions (e.g., a linear combination of the elementary indicators), while a partially compensatory (or non-compensatory) approach involves non-linear functions (e.g., a multiplicative approach).

In the first case, which corresponds to set method = "linear", the composite indicator for target unit \(c\) is obtained as weighted (according to w) arithmetic mean of the \(Q\) elementary (and normalised) indicators \(I_{qc}\):

$$CI_c = \sum_{q=1}^Q w_q I_{qc}$$

In the second case, using method = "non-linear", the resulting composite indicator is obtained as weighted geometric mean of the elementary indicators:

$$CI_c = \prod_{q=1}^Q I_{qc}^{w_q}$$

Examples

if (FALSE) {
if(interactive()){
  mock_data_core_variants <- unnest(mock_data_core, varianti, keep_empty = TRUE)
  out_companies <- ind_all(data = mock_data_core,
                          data_ind8 = mock_data_core_variants,
                          emergency_name = "coronavirus",
                          target_unit = "companies")
  data_matrix <- create_indicator_matrix(out_companies)
  data_matrix_norm <- normalise(data_matrix, method = "binary", cutoff = 0.95)
  data_matrix_norm_nomiss <- manage_missing(data_matrix_norm, missing = 0)
  w <- get_weights(data_matrix_norm_nomiss, method = "equal")
  out_aggr <- aggregate(data_matrix_norm_nomiss, method = "linear", w) #it is a composite indicator
 }
}