Compute Awarded economic value across the crisis indicator

The indicator focuses on companies that after the emergency outbreak were awarded public contracts much higher in economic value than before the emergency

Motivation:

The red flag considers at risk those companies that exceptionally increase their competitive power over the outbreak, in terms of economic value of their awarded contracts on the relevant economic market.

Scoring Rule

The output will give $1 - pvalue$, which will then be dichotomised to 1 if statistical test is significant, 0 otherwise.

Main target unit

This indicator targets Companies

Usage

ind_2(
  data,
  contract_value,
  publication_date,
  emergency_name,
  stat_unit,
  test_type,
  cpvs,
  ...
)

Arguments

data

This argument should be a data frame or tibble containing the data you want to use to calculate the indicator.

contract_value

This argument corresponds to the name of the column in data containing the overall amount of the tender for each contract. The values in this column should be numeric.

publication_date

This argument corresponds to the name of the column in data containing the publication date for each notice or report.

emergency_name

This argument should be a character string specifying the name of the emergency or event you are analyzing. Examples could include "Coronavirus" or "Terremoto Aquila".

stat_unit

This argument should be a character string specifying the statistical unit of measurement or aggregation variable for the indicator. In this indicator companies are the target.

test_type

This argument should be a character vector specifying the type of hypothesis test (belonging to category 2 i.e. see statistical_tests.R) to apply to the data. Available options are "wilcoxon" and "ks".

cpvs

character vector of macro-cpv on which data is filtered out. A panel of experts have already chosen which cpvs are most affected by which emergency for you.

...

other parameters to pass to generate_indicator_schema as country_name if that not Italy, which is default behavior.

Value

indicator schema as from generate_indicator_schema() rows determined by aggregation level and indicator_value based on statistical test performed in ind_2

See also

Examples

if (FALSE) {
if (interactive()) {
  data("mock_data_core")
  ind_2(
    data = mock_data_core,
    contract_value = importo_complessivo_gara,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    test_type = "wilcoxon",
    emergency_name = "coronavirus"
  )
}
}