Compute Contract length deviation across the crisis indicator

The indicator reveals whether there has been an increase in the deviation of the contract actual execution duration from its stated/expected duration.

Motivation

The red flag considers at risk companies/contracting authorities whose contracts undergo a significant increase of their length deviation ratio - i.e., ratio between contract actual execution duration and expected duration - across the crisis.

Scoring rule

The computation procedure returns 1 - p-value of the involved test (so that high values of the indicator correspond to high levels of corruption risk). When computing the composite, it will be dichotomised to 1 if statistical test is significant, and 0 otherwise (see normalise()).

Main target unit

This indicator targets companies and contracting authorities.

Usage

ind_4(
  data,
  exp_end,
  eff_start,
  eff_end,
  stat_unit,
  emergency_name,
  publication_date,
  test_type,
  cpvs,
  ...
)

Arguments

data

a dataframe containing the data to use for computing the indicator.

exp_end

name of the variable in data containing the expected end date of each contract, i.e., contract completion date.

eff_start

name of the variable in data containing the effective start date of each contract.

eff_end

name of the variable in data containing the contract execution effective end date.

stat_unit

name of the variable in data containing the target unit ID (in this case, company or contracting authority).

emergency_name

string specifying the name of the emergency to consider. Examples could include "Coronavirus" or "Terremoto Centro Italia 2016-2017".

publication_date

name of the variable in data containing the publication date of each contract.

test_type

string specifying the statistical test to use for computing the indicator. Available options are "wilcoxon" and "ks" (Kolmogorov-Smirnov test).

cpvs

character vector of CPV divisions (first two digits of CPV code) on which data are filtered out. Note: a panel of experts have already chosen which CPV divisions are most affected by which emergency.

...

other parameters to pass to generate_indicator_schema(), such as country_name (default: Italy).

Value

indicator schema as from generate_indicator_schema().

See also

Examples

if (FALSE) {
if (interactive()) {
  ind_4(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    exp_end = data_termine_contrattuale,
    eff_end = data_effettiva_ultimazione,
    eff_start = data_inizio_effettiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}