Compute Communication default across the crisis indicator

The indicator reveals whether - and to what extent - the duty of contracting authorities to communicate the public procurement procedure activation to the Anticorruption Authority ended with a technical failure

Motivation:

The red flag considers at risk contracting authorities who fail to accomplish their duty to communicate the public procurement procedure activation to the Anticorruption Authority

Scoring Rule

If Test-statistic significant-> 1, otherwise -> 0

Main target unit

This indicator targets contracting authorities

Usage

ind_6(data, publication_date, emergency_name, award_col, stat_unit, test_type)

Arguments

data

mock_data_core example bdncp data

publication_date

The date when the tender was published

emergency_name

emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"

award_col

column indentifying id for that contract award

stat_unit

statistical unit of measurement and stat_uniting

test_type

type of the test we would like to apply

Value

indicator schema as from generate_indicator_schema

See also

Examples

if (FALSE) {
if (interactive()) {
  data("mock_data_core")
  ind_6(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    emergency_name = "coronavirus",
    award_col = id_aggiudicazione,
    stat_unit = provincia,
    test_type = "fisher"
  )
}
}