Compute One-shot opportunistic companies over the crisiss indicator

The indicator focuses on companies that after the emergency outbreak were awarded one or more public contracts having participated in the public procurement process without winning in the 5 years before the emergency outbreak

Motivation:

The red flag considers at risk those companies that show a "one- shot opportunistic behaviour", that is, companies that after the emergency outbreak were awarded one or more public contracts but did not show any competitive power in the previous years

Scoring Rule

If a company wins one or more contracts after the emergency outbreak having participated in the public procurement process without winning in the 5 years before the emergency outbreak -> 1, otherwise 0

Main target unit

This indicator targets companies

Usage

ind_7(data, publication_date, final_award_date, emergency_name, stat_unit)

Arguments

data

mock_data_core exmaple data

publication_date

The date when the tender was published

final_award_date

Date of award, as per the minutes

emergency_name

emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"

stat_unit

The unique ID Code that identifies the awarded company (ex. VAT or Tax Number)

Value

indicator schema as from generate_indicator_schema

See also

Examples

if (FALSE) {
if (interactive()) {
  data("mock_data_core")
  ind_7(
    data = mock_data_core,
    publication_date = data_pubblicazione,
    final_award_date = data_aggiudicazione_definitiva,
    stat_unit = cf_amministrazione_appaltante,
    emergency_name = "coronavirus"
  )
}
}