Italian Banca Dati Nazionali dei Contratti Pubblici (BDNCP)

Data on public contracts in Italy are published in open format on the portal, handled by the Italian anticorruption authority (ANAC).

Usage

mock_data_core

Format

mock_data_core

A data frame with approximately 300,000 contracts and 39 variables:

cig

Contract identification code: it is a code assigned by the Authority to uniquely track tenders and contracts nationwide and is assigned at the lot level.

importo_lotto

Economic amount (starting bid).

oggetto_lotto

Object (string).

oggetto_principale_contratto

Main contract object: works ("LAVORI") , services ("SERVIZI"), supplies ("FORNITURE").

cod_cpv

Common Procurement Vocabulary (CPV) code.

descrizione_cpv

Common Procurement Vocabulary (CPV) description.

flag_prevalente

Dummy, indicating whether the CPV is prevailing

luogo_istat

Contract geographic localisation (ISTAT code).

data_pubblicazione

Publication date of the call for tenders.

cod_tipo_scelta_contraente

Code referring to the procedure type for choosing a contractor.

tipo_scelta_contraente

Description of the procedure type for selecting the contractor.

cod_modalita_realizzazione

Code referring to the mode of implementation of the contract.

modalita_realizzazione

Description of the mode of implementation of the contract.

cf_amministrazione_appaltante

ID code (codice fiscale) of the contracting authorities issuing the contract.

denominazione_amministrazione_appaltante

Name of the contracting authorities issuing the contract.

citta_codice_stazapp

Contracting authority geographic location (ISTAT code).

id_aggiudicazione

Award notice ID.

data_aggiudicazione_definitiva

Award notice date.

importo_aggiudicazione

Award value.

criterio_aggiudicazione

Award criterion (most economically advantageous bid, lowest price, etc.).

numero_offerte_ammesse

Number of eligible bids.

num_imprese_offerenti

Number of received bids.

data_termine_contrattuale

Expected end date.

data_inizio_effettiva

Actual start date.

data_effettiva_ultimazione

Actual end date.

aggiudicatari

Nested dataframe containing basic information on the awarded company(ies): codice_fiscale (Company ID) and denominazione (company denomination).

varianti

Nested dataframe containing information on the modification(s) occured: id_variante(variant ID), data_approvazione_variante (variant date), cod_motivo_variante (variant reason code), and motivo_variante (variant reason description).

cod_regione

Region ISTAT code (NUTS-2 level).

nome_regione

Region name (NUTS-2 level).

cod_provincia

Province ISTAT code (NUTS-3 level).

nome_provincia

Province name (NUTS-3 level).

cod_comune

Municipality ISTAT code (local administrative units).

nome_comune

Municipality name (local administrative units).

rip_geo

Geographic macro-level (NUTS-1 level).

codice_nuts1_2021

NUTS-1 code.

codice_nuts2_2021

NUTS-2 code.

codice_nuts3_2021

NUTS-3 code.

Source

https://dati.anticorruzione.it/opendata

Details

Nested dataframes (see aggiudicatari and varianti) can be unnested using function tidyr::unnest().