Compute Awarded economic value across the crisis indicator

The indicator focuses on companies that after the emergency outbreak were awarded public contracts much higher in economic value than before the emergency

Motivation:

The red flag considers at risk those companies that exceptionally increase their competitive power over the outbreak, in terms of economic value of their awarded contracts on the relevant economic market.

Scoring Rule

If Test-statistic significant-> 1, otherwise -> 0

Main target unit

This indicator targets Companies

Usage

ind_2(
  data,
  contract_value,
  publication_date,
  emergency_name,
  stat_unit,
  test_type
)

Arguments

data

test bndcp data

contract_value

Overall Amount of the Tender

publication_date

the date in which the tender was published

emergency_name

emergency name character string for which you want to evaluate the indicator, e.g. "Coronavirus" "Terremoto Aquila"

stat_unit

statistical target unit of measurement, aggregation variable, the indicator target

test_type

character vector identifying the type of test you want to execute, alternatives are c("ks", "wilcoxon")

Value

indicator schema as from generate_indicator_schema() rows determined by aggregation level and indicator_value based on statistical test performed in ind_2

See also

Examples

if (FALSE) {
if (interactive()) {
  data("mock_data_core")
  ind_2(
    data = mock_data_core,
    contract_value = importo_complessivo_gara,
    publication_date = data_pubblicazione,
    stat_unit = provincia,
    test_type = "ks",
    emergency_name = "coronavirus"
  )
}
}