BDNCP open data from ANAC

Data on public contracts in Italy are published in open open format on the portal. The data are those communicated in their original form to the Authority by the Single Procedure Managers of the Contracting Stations. In this case we merged the tables whose dimensions i.e. columns if joined to "bando cig", are to calculate indicators.

Usage

mock_data_core

Format

mock_data_core

A data frame with 7,240 rows and 60 columns:

cig, cig_accordo_quadro

Tender Identification Code, is a code assigned by the Authority to uniquely track tenders and contracts nationwide and is assigned at the Lot level.

numero_gara,

number identifying a tender; a tender consists of one or more lots, each identified by a CIG

oggetto_gara

Subject of the Tender - description of goods and services to be procured

importo_complessivo_gara

Overall Amount of the Tender

n_lotti_componenti

Number of lots that make up the contract

oggetto_lotto

Subject of the Lot

importo_lotto

Lot amount

oggetto_principale_contratto

Main object of the contract (Works, Services Supplies)

stato

cig life cycle status

settore

indicates whether the Contracting Authority operates in the ordinary sectors (EU Directive 24 applies) or in the Special Sectors (EU Directive 25 applies)

luogo_istat

ISTAT Localization

provincia

Province abbreviation

data_pubblicazione

date of publication of the notice

data_scadenza_offerta

Deadline for submission of a bid

cod_tipo_scelta_contraente

Code referring to the procedure for choosing a contractor

tipo_scelta_contraente

Description of the procedure for selecting the contractor

cod_modalita_realizzazione

Code referring to the mode of implementation of the tender

modalita_realizzazione

Description of the mode of implementation of the tender

codice_ausa

AUSA Code (a code assigned by ANAC (italian national anti-corruption authority), which identifies the Contracting Station i.e. authority in the Unique Register of Contracting Stations)

cf_amministrazione_appaltante

Fiscal Code of the Administration that tenders or purchases

denominazione_amministrazione_appaltante

Name of the administration issuing the invitation to tender or purchase

sezione_regionale

Regional reference section for data submission.

id_centro_costo

Identifier of the Cost Center, organizational unit of the contracting station.

denominazione_centro_costo

Cost Center Name

anno_pubblicazione

Reference year, extracted from the date of publication

mese_pubblicazione

Reference month, extracted from the date of publication

cod_cpv

Commodity category code of the purchased good or service (CPV vocabulary)

descrizione_cpv

Description of the commodity category of the good or service purchased

flag_prevalente

denotes whether the commodity category is prevailing (a contract may include more than one category of goods or services)

cup

Unique Project Code, identifies a public investment project

id_categoria

Identification code for the category of works being procured

descrizione

Description of the categories of works

descrizione_tipo_categoria

Description of the type of work category of the contract

modalita_realizzazione_denominazione

Name of the mode of implementation of the contract

tipo_scelta_contraente_denominazione

name of the Contractor's Choice

data_aggiudicazione_definitiva

Date of award, as per the minutes

esito

Description of the outcome of the award process (Awarded, Cancelled...)

criterio_aggiudicazione

Award criterion adopted (most economically advantageous bid, lowest price, etc.)

data_comunicazione_esito

Date of announcement of the outcome of the award.

importo_aggiudicazione

amount of the award

ribasso_aggiudicazione

Value of the rebate made by the Economic Operator who was awarded the tender.

flag_subappalto

Indicator of possibility of subcontracting

id_aggiudicazione

Unique award code

cod_esito

Code identifying outcome of award procedure

asta_elettronica

Indicator of use of electronic auction

massimo_ribasso

Maximum value of the discount charged

minimo_ribasso

Minimum value of the discount charged in the tender

data

reference date for the data block

descrizione_evento

Event Description (Auction Base, Variant...)

importo_sicurezza

Event Description (Auction Base, Variant...)

dettaglio_evento

Detail of variation e.g. variant number

importo_forniture

Amount of Supplies

importo_lavori

Amount of Work

importo_progettazione

Design Amount

somme_a_disposizione

Amounts available

importo_servizi

Amount of Services

ulteriori_oneri_non_soggetti_ribasso

Additional charges not subject to rebate

ruolo

Description of the participant's role.

codice_fiscale

Participant's tax code.

denominazione

Participant name.

tipo_soggetto

Type of participating entity.

cf_subappaltante

Subcontractor Tax Code.

data_autorizzazione

Subcontracting authorization date

oggetto

Subject of subcontracting.

descrizione_categoria

Description of the category of works.

cod_categoria

Works category identification code.

data_stipula_contratto

Date of entering into the contract

data_esecutivita_contratto

Contract execution date

data_termine_contrattuale

Contract termination date

data_inizio_effettiva

Actual start date of work

consegna_frazionata

Fractional delivery indicator

id_variante

Unique variant identifier.

motivo_variante

Description of the reason for the variance.

data_approvazione_variante

Date of approval of the variance.

motivo_risoluzione

Reason for contract resolution

motivo_interruzione_anticipata

Description of the reason for early termination

data_conclusione_anticipata

Date of early termination of the contract

data_effettiva_ultimazione

Effective completion date of the contract

Source

https://dati.anticorruzione.it/opendata